--- src/main/java/net/ccbluex/liquidbounce/features/module/modules/exploit/phasemodes/aac/AAC350.kt
@@ -1,38 +1,0 @@
-/*
- * LiquidBounce Hacked Client
- * A free open source mixin-based injection hacked client for Minecraft using Minecraft Forge.
- * https://github.com/CCBlueX/LiquidBounce/
- */
-package net.ccbluex.liquidbounce.features.module.modules.exploit.phasemodes.aac
-
-import net.ccbluex.liquidbounce.event.UpdateEvent
-import net.ccbluex.liquidbounce.event.PacketEvent
-import net.ccbluex.liquidbounce.features.module.modules.exploit.Phase.tickTimer
-import net.ccbluex.liquidbounce.features.module.modules.exploit.phasemodes.PhaseMode
-import net.minecraft.network.play.client.C03PacketPlayer
-
-object AAC350 : PhaseMode("AAC3.5.0") {
-	private var tick = 0
-	
-	override fun onUpdate(event: UpdateEvent) {
-        if (tickTimer.hasTimePassed(2) && mc.thePlayer.isCollidedHorizontally && (!isInsideBlock || mc.thePlayer.isSneaking)) {
-            val yaw = Math.toRadians(mc.thePlayer.rotationYaw.toDouble())
-            val oldX = mc.thePlayer.posX
-            val oldZ = mc.thePlayer.posZ
-            val x = -sin(yaw)
-            val z = cos(yaw)
-
-            mc.thePlayer.setPosition(oldX + x, mc.thePlayer.posY, oldZ + z)
-            tickTimer.reset()
-        }
-	}
-
-    override fun onPacket(event: PacketEvent) {
-        if (event.packet is C03PacketPlayer) {
-            val yaw = direction.toFloat()
-
-            event.packet.x -= sin(yaw.toDouble()) * 0.00000001
-            event.packet.z += cos(yaw.toDouble()) * 0.00000001
-        }
-    }
-}
